new coordinates

+++ include/structures.h
@@ -6,7 +6,7 @@
 /*   By: saguesse <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 14:59:31 by saguesse          #+#    #+#             */
-/*   Updated: 2023/02/28 16:19:54 by saguesse         ###   ########.fr       */
+/*   Updated: 2023/03/01 15:34:24 by saguesse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ typedef struct s_img
 typedef struct s_ambient
 {
 	char	*identifier;
-	float	ratio;
+	double	ratio;
 	int		r;
 	int		g;
 	int		b;
@@ -37,22 +37,22 @@ typedef struct s_ambient
 typedef struct s_camera
 {
 	char	*identifier;
-	float	x;
-	float	y;
-	float	z;
-	float	v_x;
-	float	v_y;
-	float	v_z;
+	double	x;
+	double	y;
+	double	z;
+	double	v_x;
+	double	v_y;
+	double	v_z;
 	int		fov;
 }	t_camera;
 
 typedef struct s_light
 {
 	char	*identifier;
-	float	x;
-	float	y;
-	float	z;
-	float	light;
+	double	x;
+	double	y;
+	double	z;
+	double	light;
 }	t_light;
 
 typedef struct s_data
